requirements.txt
numpy==2.4.1
Pillow==12.1.0
streamlit==1.45.1
torch==2.7.0
torchvision==0.22.0
